mcp-sse-server/pyproject.toml
[project]
name = "mcp-sse-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-aws>=0.2.24",
    "langchain-community>=0.3.24",
    "langchain-mcp-adapters>=0.1.4",
    "langgraph>=0.4.7",
    "mcp[cli]>=1.9.2",
    "python-dotenv>=1.1.0",
]
